extract tests in respective files
